Remplacé les types 'long' par des 'int' pour les egids et les ewids.
git-svn-id: https://svn.etat-de-vaud.ch/SVN_SOCLE/registre/unireg/trunk@55713 cb3efa0a-b8b0-9f43-a4c8-343d8b7e4362

+++ unireg/business/src/main/java/ch/vd/uniregctb/interfaces/model/Adresse.java
@@ -106,10 +106,10 @@ public interface Adresse extends DateRange, AdresseAvecCommune {
 	/**
 	 * @return le numéro Ofs de bâtiment (Gebäude) ou <b>null</b> s'il est inconnu.
 	 */
-	Long getEgid();
+	Integer getEgid();
 
 	/**
 	 * @return le numéro Ofs de logement (Wohnung) ou <b>null</b> s'il est inconnu.
 	 */
-	Long getEwid();
+	Integer getEwid();
 }
